requirements.txt
altgraph==0.17.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
dj-database-url==2.1.0
Django==5.1.7
django-cors-headers==4.3.1
django-minio-storage==0.5.7
django-storages==1.14.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.1
drf-yasg==1.21.7
gunicorn==23.0.0
inflection==0.5.1
macholib==1.16.3
minio==7.2.15
packaging==24.2
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.22.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
setuptools==74.1.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
